✅ TFBJEG code: split context profile switch helpers
